Some warnings and an example for clarity
Added a warning about overlapping networks for the LAN and Wireguard interfaces, as doing so will cause the clients not to be able to communicate with each other properly. Added an example to request another PD from your ISP, as ATT only hands out /64s and one will not be enough to subnet the networks properly.