Hardware / BIOS: add newest A20 BIOS
This commit provides two new images for the Netboard A20 BIOS.

Note: the flashtool firmware has changed in such a way that installing
the latest BIOS (version 10) will prevent a user from reverting to
version 9. Therefore two new images for the old version 9 are also
updated here so as to not break backwards compatibility. The checksums
have been updated accordingly