manual: user must 'Apply' changes after generating wg peer (#634)
* manual/vpnet: user must apply changes after generating wg peer

After clicking "Store and generate next," `wg showconf wg1` on opnsense doesn't show the new peer (and connection attempts from the peer fail without explanation). Visiting the WG->Peers page and clicking "Apply" fixes this.

* manual/how-tos/wireguard-client: s/Save/Apply/ after saving a peer or instance

Perhaps the wording in the webui changed, or it was just an oversight in the original doc, but the button on the WG->Peers and WG->Instances pages is labeled "Apply" (as of 24.7).